Version 1.2- Added modified files to src folder
Handled the number buttons onClick as well as the math operators '+',' -', '/', '*'.  Handled use case of user pressing the operator button twice.  The second consecutive time the operator is pressed will not be added to the display's state.  Also, the operator will not be added if the display state is empty.  I wanted to learn how to play a sound onClick, so I have added a sound file, audio constructor, and function to play a sound when the user presses a button.  I still need to handle the text within the virtual display screen expanding beyond the boundaries of the div in which it resides.  I plan to handle the submit ('=') button and clear button next.  This will leave the percentage sign and the parentheses buttons